Remove readarr type file (#49)
