+++ readarr/readarr.go
@@ -1,13 +1,14 @@
 package readarr
 
 import (
-	"context"
 	"crypto/tls"
-	"fmt"
 	"net/http"
 
 	"golift.io/starr"
 )
+
+// APIver is the Readarr API version supported by this library.
+const APIver = "v1"
 
 // Readarr contains all the methods to interact with a Readarr server.
 type Readarr struct {
@@ -51,96 +52,4 @@ func New(config *starr.Config) *Readarr {
 	}
 
 	return &Readarr{APIer: config}
-}
-
-// GetQueue returns a single page from the Readarr Queue (processing, but not yet imported).
-// WARNING: 12/30/2021 - this method changed.
-// If you need control over the page, use readarr.GetQueuePage().
-// This function simply returns the number of queue records desired,
-// up to the number of records present in the application.
-// It grabs records in (paginated) batches of perPage, and concatenates
-// them into one list.  Passing zero for records will return all of them.
-func (r *Readarr) GetQueue(records, perPage int) (*Queue, error) {
-	return r.GetQueueContext(context.Background(), records, perPage)
-}
-
-func (r *Readarr) GetQueueContext(ctx context.Context, records, perPage int) (*Queue, error) {
-	queue := &Queue{Records: []*QueueRecord{}}
-	perPage = starr.SetPerPage(records, perPage)
-
-	for page := 1; ; page++ {
-		curr, err := r.GetQueuePageContext(ctx, &starr.Req{PageSize: perPage, Page: page})
-		if err != nil {
-			return nil, err
-		}
-
-		queue.Records = append(queue.Records, curr.Records...)
-
-		if len(queue.Records) >= curr.TotalRecords ||
-			(len(queue.Records) >= records && records != 0) ||
-			len(curr.Records) == 0 {
-			queue.PageSize = curr.TotalRecords
-			queue.TotalRecords = curr.TotalRecords
-			queue.SortDirection = curr.SortDirection
-			queue.SortKey = curr.SortKey
-
-			break
-		}
-
-		perPage = starr.AdjustPerPage(records, curr.TotalRecords, len(queue.Records), perPage)
-	}
-
-	return queue, nil
-}
-
-// GetQueuePage returns a single page from the Readarr Queue.
-// The page size and number is configurable with the input request parameters.
-func (r *Readarr) GetQueuePage(params *starr.Req) (*Queue, error) {
-	return r.GetQueuePageContext(context.Background(), params)
-}
-
-func (r *Readarr) GetQueuePageContext(ctx context.Context, params *starr.Req) (*Queue, error) {
-	var queue Queue
-
-	params.CheckSet("sortKey", "timeleft")
-	params.CheckSet("includeUnknownAuthorItems", "true")
-
-	_, err := r.GetInto(ctx, "v1/queue", params.Params(), &queue)
-	if err != nil {
-		return nil, fmt.Errorf("api.Get(queue): %w", err)
-	}
-
-	return &queue, nil
-}
-
-// GetRootFolders returns all configured root folders.
-func (r *Readarr) GetRootFolders() ([]*RootFolder, error) {
-	return r.GetRootFoldersContext(context.Background())
-}
-
-func (r *Readarr) GetRootFoldersContext(ctx context.Context) ([]*RootFolder, error) {
-	var folders []*RootFolder
-
-	_, err := r.GetInto(ctx, "v1/rootFolder", nil, &folders)
-	if err != nil {
-		return nil, fmt.Errorf("api.Get(rootFolder): %w", err)
-	}
-
-	return folders, nil
-}
-
-// GetMetadataProfiles returns the metadata profiles.
-func (r *Readarr) GetMetadataProfiles() ([]*MetadataProfile, error) {
-	return r.GetMetadataProfilesContext(context.Background())
-}
-
-func (r *Readarr) GetMetadataProfilesContext(ctx context.Context) ([]*MetadataProfile, error) {
-	var profiles []*MetadataProfile
-
-	_, err := r.GetInto(ctx, "v1/metadataprofile", nil, &profiles)
-	if err != nil {
-		return nil, fmt.Errorf("api.Get(metadataprofile): %w", err)
-	}
-
-	return profiles, nil
 }
