+++ readarr/system.go
@@ -3,9 +3,42 @@ package readarr
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"golift.io/starr"
 )
+
+// SystemStatus is the /api/v1/system/status endpoint.
+type SystemStatus struct {
+	Version                string    `json:"version"`
+	BuildTime              time.Time `json:"buildTime"`
+	IsDebug                bool      `json:"isDebug"`
+	IsProduction           bool      `json:"isProduction"`
+	IsAdmin                bool      `json:"isAdmin"`
+	IsUserInteractive      bool      `json:"isUserInteractive"`
+	StartupPath            string    `json:"startupPath"`
+	AppData                string    `json:"appData"`
+	OsName                 string    `json:"osName"`
+	OsVersion              string    `json:"osVersion"`
+	IsNetCore              bool      `json:"isNetCore"`
+	IsMono                 bool      `json:"isMono"`
+	IsLinux                bool      `json:"isLinux"`
+	IsOsx                  bool      `json:"isOsx"`
+	IsWindows              bool      `json:"isWindows"`
+	IsDocker               bool      `json:"isDocker"`
+	Mode                   string    `json:"mode"`
+	Branch                 string    `json:"branch"`
+	Authentication         string    `json:"authentication"`
+	SqliteVersion          string    `json:"sqliteVersion"`
+	MigrationVersion       int       `json:"migrationVersion"`
+	URLBase                string    `json:"urlBase"`
+	RuntimeVersion         string    `json:"runtimeVersion"`
+	RuntimeName            string    `json:"runtimeName"`
+	StartTime              time.Time `json:"startTime"`
+	PackageVersion         string    `json:"packageVersion"`
+	PackageAuthor          string    `json:"packageAuthor"`
+	PackageUpdateMechanism string    `json:"packageUpdateMechanism"`
+}
 
 // GetSystemStatus returns system status.
 func (r *Readarr) GetSystemStatus() (*SystemStatus, error) {
